pyproject.toml
[project]
name = "ai-driven-development-project-template"
version = "0.1.0"
description = "AWS サーバーレスマイクロサービス プロジェクトテンプレート"
requires-python = ">=3.13"
dependencies = [
    "aws-lambda-powertools>=3.0.0",
    "pydantic[email]>=2.0.0",
    "boto3>=1.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "moto[dynamodb,events]>=5.0.0",
    "requests>=2.32.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "boto3-stubs[dynamodb,events]>=1.35.0",
]
cdk = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.0.0",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "ANN",  # flake8-annotations
    "S",    # flake8-bandit (security)
    "T20",  # flake8-print
    "RET",  # flake8-return
    "ARG",  # flake8-unused-arguments
]
ignore = [
    "ANN101",  # Missing type annotation for self
    "ANN102",  # Missing type annotation for cls
    "ANN401",  # Dynamically typed expressions (typing.Any)
]

[tool.ruff.lint.isort]
known-first-party = ["services"]

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = ["S101", "ANN"]  # assert と型注釈をテストでは許可

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "moto",
    "moto.*",
    "aws_lambda_powertools.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["services", "tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
